VERSION 2.8 - Les photos prises avec l'appareil photo sont maintenant stockées en haute définition - L'import de photos depuis la galerie est maintenant possible - Ajout de pays (top 20 des pays producteurs de vin dans le monde) - Correction de bugs - Amélioration de l'affichage

+++ src/fr/kougteam/myCellar/activity/AccordMetVinActivity.java
@@ -140,7 +140,7 @@ public class AccordMetVinActivity extends Activity {
 		vinsCellierListView.setOnItemClickListener(new OnItemClickListener() {		 
 		    public void onItemClick(AdapterView<?> parent, View view, int pos, long id) {
 		    	Cursor selectedItem = (Cursor)parent.getItemAtPosition(pos);
-		    	int idVin = selectedItem.getInt(selectedItem.getColumnIndex(VinDao.COL_ID));
+		    	long idVin = selectedItem.getLong(selectedItem.getColumnIndex(VinDao.COL_ID));
 		    	intent2DetailVin.putExtra("idVin", idVin);
 				startActivity(intent2DetailVin);
 			}
