+++ src/fr/kougteam/myCellar/modele/Met.java
@@ -3,7 +3,7 @@ package fr.kougteam.myCellar.modele;
 import java.io.Serializable;
 
 /**
- * Représente un met
+ * Reprï¿½sente un met
  * 
  * @author Thomas Cousin
  *
@@ -11,22 +11,22 @@ import java.io.Serializable;
 public class Met implements Serializable {
 
 	private static final long serialVersionUID = 3428679995842594922L;
-	private int id;
+	private long id;
 	private String nom;
 		
 	public Met() {
 		super();
 	}
-	public Met(int id, String nom) {
+	public Met(long id, String nom) {
 		super();
 		this.id = id;
 		this.nom = nom;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNom() {
