+++ src/fr/kougteam/myCellar/modele/MetVin.java
@@ -3,7 +3,7 @@ package fr.kougteam.myCellar.modele;
 import java.io.Serializable;
 
 /**
- * Représente une association met / vin
+ * Reprï¿½sente une association met / vin
  * 
  * @author Thomas Cousin
  *
@@ -11,23 +11,23 @@ import java.io.Serializable;
 public class MetVin implements Serializable {
 
 	private static final long serialVersionUID = 3428679995842594922L;
-	private int idMet;
+	private long idMet;
 	private String nomVin;
 	private String type;
 		
 	public MetVin() {
 		super();
 	}
-	public MetVin(int idMet, String nomVin, String type) {
+	public MetVin(long idMet, String nomVin, String type) {
 		super();
 		this.idMet = idMet;
 		this.nomVin = nomVin;
 		this.type = type;
 	}
-	public int getIdMet() {
+	public long getIdMet() {
 		return idMet;
 	}
-	public void setIdMet(int idMet) {
+	public void setIdMet(long idMet) {
 		this.idMet = idMet;
 	}
 	public String getNomVin() {
