+++ src/fr/kougteam/myCellar/modele/Pays.java
@@ -3,7 +3,7 @@ package fr.kougteam.myCellar.modele;
 import java.io.Serializable;
 
 /**
- * Représente un pays
+ * Reprï¿½sente un pays
  * 
  * @author Thomas Cousin
  *
@@ -11,22 +11,22 @@ import java.io.Serializable;
 public class Pays implements Serializable {
 
 	private static final long serialVersionUID = -4336017170678429293L;
-	private int id;
+	private long id;
 	private String nom;
 		
 	public Pays() {
 		super();
 	}
-	public Pays(int id, String nom) {
+	public Pays(long id, String nom) {
 		super();
 		this.id = id;
 		this.nom = nom;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNom() {
